bump to 4o-mini, auto gen if > 90%, reduce target codebook size, clean up
